Made it work and for linux
the image wouldnt load previous, I fixed that and made it accessible for linux CONSOLE ONLY. 
If you try to run this as just a linux desktop executable, you will be confused. 

Fixed naming and spacing of coordinates generated.
problem remains: need a line break after every 4 pairs of (x,y) coordinates and before ever spot ID. 
We also need to fix the UI in the app, have the spots begin at "1" instead of at  "0" would be more practical for users.